+++ Terraria/GameContent/Biomes/Desert/SurfaceMap.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.GameContent.Biomes.Desert.SurfaceMap
-// Assembly: Terraria, Version=1.4.3.6, Culture=neutral, PublicKeyToken=null
-// MVID: F541F3E5-89DE-4E5D-868F-1B56DAAB46B2
+// Assembly: Terraria, Version=1.4.4.9, Culture=neutral, PublicKeyToken=null
+// MVID: CD1A926A-5330-4A76-ABC1-173FBEBCC76B
 // Assembly location: D:\Program Files\Steam\steamapps\content\app_105600\depot_105601\Terraria.exe
 
 using System;
@@ -10,7 +10,7 @@ namespace Terraria.GameContent.Biomes.Desert
 {
   public class SurfaceMap
   {
-    public readonly float Average;
+    public readonly double Average;
     public readonly int Bottom;
     public readonly int Top;
     public readonly int X;
@@ -35,7 +35,7 @@ namespace Terraria.GameContent.Biomes.Desert
         val1_1 = (int) Main.worldSurface - 10;
       this.Bottom = val1_1;
       this.Top = val1_2;
-      this.Average = (float) num / (float) this._heights.Length;
+      this.Average = (double) num / (double) this._heights.Length;
     }
 
     public short this[int absoluteX] => this._heights[absoluteX - this.X];
