+++ Terraria/GameContent/Generation/PassLegacy.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.GameContent.Generation.PassLegacy
-// Assembly: Terraria, Version=1.4.3.6, Culture=neutral, PublicKeyToken=null
-// MVID: F541F3E5-89DE-4E5D-868F-1B56DAAB46B2
+// Assembly: Terraria, Version=1.4.4.9, Culture=neutral, PublicKeyToken=null
+// MVID: CD1A926A-5330-4A76-ABC1-173FBEBCC76B
 // Assembly location: D:\Program Files\Steam\steamapps\content\app_105600\depot_105601\Terraria.exe
 
 using System.Collections.Generic;
@@ -801,12 +801,12 @@ namespace Terraria.GameContent.Generation
     private readonly WorldGenLegacyMethod _method;
 
     public PassLegacy(string name, WorldGenLegacyMethod method)
-      : base(name, PassLegacy.GetWeight(name))
+      : base(name, (double) PassLegacy.GetWeight(name))
     {
       this._method = method;
     }
 
-    public PassLegacy(string name, WorldGenLegacyMethod method, float weight)
+    public PassLegacy(string name, WorldGenLegacyMethod method, double weight)
       : base(name, weight)
     {
       this._method = method;
