+++ Terraria/WorldBuilding/Passes.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.WorldBuilding.Passes
-// Assembly: Terraria, Version=1.4.3.6, Culture=neutral, PublicKeyToken=null
-// MVID: F541F3E5-89DE-4E5D-868F-1B56DAAB46B2
+// Assembly: Terraria, Version=1.4.4.9, Culture=neutral, PublicKeyToken=null
+// MVID: CD1A926A-5330-4A76-ABC1-173FBEBCC76B
 // Assembly location: D:\Program Files\Steam\steamapps\content\app_105600\depot_105601\Terraria.exe
 
 using Terraria.IO;
@@ -13,7 +13,7 @@ namespace Terraria.WorldBuilding
     public class Clear : GenPass
     {
       public Clear()
-        : base("clear", 1f)
+        : base("clear", 1.0)
       {
       }
 
@@ -41,7 +41,7 @@ namespace Terraria.WorldBuilding
 
       public ScatterCustom(
         string name,
-        float loadWeight,
+        double loadWeight,
         int count,
         GenBase.CustomPerUnitAction perUnit = null)
         : base(name, loadWeight)
